Frontend folder reorganization
