+++ client/src/menus/Login.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import './App.css'
+import './../App.css'
 
 const Login = () => {
   const [roomCode, setRoomCode] = useState('');
