+++ client/src/menus/RoomFull.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from 'react-router-dom';
-import './App.css'
+import './../App.css'
 
 const RoomFull = () => {
   const navigate = useNavigate();
