+++ client/src/tests/Login.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import Login from '../Login';
+import Login from '../menus/Login';
 import '@testing-library/jest-dom';
 import { vi } from 'vitest';
