+++ client/src/tests/RoomFull.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import RoomFull from '../RoomFull';
+import RoomFull from '../menus/RoomFull';
 import '@testing-library/jest-dom';
 import { vi } from 'vitest';
